+++ netgore/trunk/DemoGame.ServerObjs/Queries/Character/Items/SelectCharacterEquippedItemsQuery.cs
@@ -41,7 +41,7 @@ namespace DemoGame.Server.Queries
 
         protected override void SetParameters(DbParameterValues p, CharacterID characterID)
         {
-            p["@characterID"] = characterID;
+            p["@characterID"] = (int)characterID;
         }
     }
 }
