+++ netgore/trunk/DemoGame.ServerObjs/Queries/Character/SelectCharacterByIDQuery.cs
@@ -38,7 +38,7 @@ namespace DemoGame.Server.Queries
 
         protected override void SetParameters(DbParameterValues p, CharacterID characterID)
         {
-            p["@characterID"] = characterID;
+            p["@characterID"] = (int)characterID;
         }
     }
 }
