+++ netgore/trunk/DemoGame.ServerObjs/Queries/Character/Template/SelectCharacterTemplateEquippedQuery.cs
@@ -43,7 +43,7 @@ namespace DemoGame.Server.Queries
 
         protected override void SetParameters(DbParameterValues p, CharacterTemplateID id)
         {
-            p["@characterID"] = id;
+            p["@characterID"] = (int)id;
         }
     }
 }
