+++ netgore/trunk/DemoGame.ServerObjs/Queries/Item/SelectItemQuery.cs
@@ -40,7 +40,7 @@ namespace DemoGame.Server.Queries
 
         protected override void SetParameters(DbParameterValues p, ItemID id)
         {
-            p["@id"] = id;
+            p["@id"] = (int)id;
         }
     }
 }
