+++ netgore/trunk/DemoGame.ServerObjs/Queries/Item/UpdateItemFieldQuery.cs
@@ -70,7 +70,7 @@ namespace DemoGame.Server.Queries
 
             protected override void SetParameters(DbParameterValues p, QueryArgs item)
             {
-                p["@itemID"] = item.ItemID;
+                p["@itemID"] = (int)item.ItemID;
                 p["@value"] = item.Value;
             }
         }
